requirements.txt
fastapi==0.115.1
uvicorn==0.23.2
httpx==0.25.0
pyjwt==2.9.0
python-dotenv==1.0.0
itsdangerous==2.2.0
anyio==3.7.1
pydantic-settings==2.12.0
requests==2.31.0
cryptography